pyproject.toml
[project]
name = "langchain-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.13.0",
    "langchain-community>=0.4.1",
    "langchain[openai]>=1.0.8",
    "python-dotenv>=1.2.1",
]
